requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
httpx
openai
streamlit
tenacity